CC/requirements.txt
Django>=4.2,<5.0
django-environ>=0.11.0
Pillow>=10.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
openai>=1.0.0
google-generativeai>=0.3.0
